feature: The keep constraint statement is analyzed preliminarily
